requirements.txt
langchain
langchain-google-genai
transformers
tensorflow==2.16.1
tf-keras  
langchain_community
python-dotenv
google-api-python-client